+++ .github/workflows/release-notes.py
@@ -28,13 +28,30 @@ Check out the [rqlite Docker page](https://hub.docker.com/r/rqlite/rqlite/) for 
 ### Linux
 _Builds for a variety of CPU architectures are available. See the Assets section below._
 
-To download and start rqlite, execute the following in a shell.
-
+The easiest way to install rqlite on Linux:
 ```
-curl -L https://github.com/rqlite/rqlite/releases/download/{release}/rqlite-{release}-linux-amd64.tar.gz -o rqlite-{release}-linux-amd64.tar.gz
-tar xvfz rqlite-{release}-linux-amd64.tar.gz
-cd rqlite-{release}-linux-amd64
-./rqlited data/
+curl -fsSL https://rqlite.io/install.sh -o install.sh
+sh install.sh
+```
+To install a specific version:
+```
+curl -fsSL https://rqlite.io/install.sh -o install.sh
+RQLITE_VERSION={release} INSTALL_DIR={location} sh install.sh
+```
+
+### Packages
+_.deb and .rpm packages for amd64 and arm64 are available in the Assets section below._
+
+Install on Debian/Ubuntu:
+```
+sudo dpkg -i rqlite_{version}_amd64.deb
+sudo systemctl start rqlited
+```
+
+Install on RHEL/Fedora:
+```
+sudo rpm -i rqlite-{version}-1.x86_64.rpm
+sudo systemctl start rqlited
 ```
 
 ### macOS
@@ -70,7 +87,8 @@ def main():
     token = sys.argv[1]
     release_id = sys.argv[2]
     tag = sys.argv[3]
-    set_release_notes(token, release_id, release_template.format(release=tag))
+    version = tag.lstrip('v')
+    set_release_notes(token, release_id, release_template.format(release=tag, version=version))
 
 if __name__ == "__main__":
     main()